Fixed problem that caused time conversion warnings during compilation.

+++ src/include/DStoreVec_impl.h
@@ -259,7 +259,7 @@ namespace Loci {
     static_storeVec.allocate(tmp_dom) ;
     static_storeVec.setVecSize(size) ;
     FORALL(tmp_dom, ei) {
-      for(size_t i = 0; i < size; ++i)
+      for(int i = 0; i < size; ++i)
 	static_storeVec[ei][i] = attrib_data[ei][i] ;
     }ENDFORALL ;
     return static_storeVec.Rep() ;
@@ -391,7 +391,7 @@ namespace Loci {
       iter = attrib_data.find(*ci);
       if( iter != attrib_data.end() ) {
         newVec = iter->second;
-        for( size_t ivec = 0; ivec < size; ivec++){
+        for( int ivec = 0; ivec < size; ivec++){
           typename schema_traits::Converter_Type cvtr( newVec[ivec] );
           arraySize += cvtr.getSize();
         }
@@ -480,7 +480,7 @@ namespace Loci {
       iter = attrib_data.find(*ci);
       if( iter == attrib_data.end() ) continue;
       newVec = iter->second;
-      for( size_t ivec = 0; ivec < size; ivec++){
+      for( int ivec = 0; ivec < size; ivec++){
           typename schema_traits::Converter_Type cvtr( newVec[ivec] );
 
           stateSize  = cvtr.getSize();
@@ -508,7 +508,7 @@ namespace Loci {
 
     int M ;
     MPI_Unpack(inbuf, insize, &position, &M, 1, MPI_INT, MPI_COMM_WORLD) ;
-    if( size_t(M) > size) {
+    if( M > size) {
       set_elem_size(M) ;
     }
     unpackdata( traits_type, inbuf, position, insize, seq);
@@ -530,7 +530,7 @@ namespace Loci {
       outcount = size*sizeof(T);
       MPI_Unpack( inbuf, insize, &position, &outbuf[0], outcount, 
                   MPI_BYTE, MPI_COMM_WORLD) ;
-      for( size_t i = 0; i < size; i++) 
+      for( int i = 0; i < size; i++) 
         attrib_data[*ci][i] = outbuf[i];
     }
 
@@ -559,7 +559,7 @@ namespace Loci {
       if( attrib_data[*ci].size() < size) {
         attrib_data[*ci].resize(size);
       }
-      for( size_t ivec = 0; ivec < size; ivec++) {
+      for( int ivec = 0; ivec < size; ivec++) {
         outcount = sizeof(int);
         MPI_Unpack(inbuf, insize, &position, &stateSize, 1, MPI_INT, 
                    MPI_COMM_WORLD) ;
